+++ result_scrapping.html
@@ -1,0 +1,1 @@
+<html><head><title>etsy.com</title><style>#cmsg{animation: A 1.5s;}@keyframes A{0%{opacity:0;}99%{opacity:0;}100%{opacity:1;}}</style></head><body style="margin:0"><p id="cmsg">Please enable JS and disable any ad blocker</p><script data-cfasync="false">var dd={'rt':'c','cid':'AHrlqAAAAAMAJI-fEUgXYocAemojCg==','hsh':'D013AA612AB2224D03B2318D0F5B19','t':'fe','s':45977,'e':'61710651ce68e7f85262d5fa0b4c44256e2060d9e150eefd5d77b7b1c15dd3f6','host':'geo.captcha-delivery.com'}</script><script data-cfasync="false" src="https://ct.captcha-delivery.com/c.js"></script></body></html>
